Move unit test classes into ".test" sub-packages and remove import of host bundle packages.
