+++ net.solarnetwork.node.datum.egauge.ws.test/META-INF/MANIFEST.MF
@@ -7,7 +7,5 @@ Bundle-Vendor: SolarNetwork
 Fragment-Host: net.solarnetwork.node.datum.egauge.ws;bundle-version="1.0.0"
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Import-Package: 
- net.solarnetwork.node.datum.egauge.ws,
- net.solarnetwork.node.datum.egauge.ws.client,
  org.junit;version="[4.5,5.0)",
  org.junit.runner;version="[4.5,5.0)"
