+++ net.solarnetwork.node.datum.egauge.ws.test/src/net/solarnetwork/node/datum/egauge/ws/test/client/test/instantaneous.xml
